refactor: optimization algorithms
We will delegate the definition of the loss function to the classes
defined in the lossFunction package. We also move the optimization
algorithms to the algorithms package.
